split out graph component
