+++ src/components/Graph/graphHelpers.js
